Correções de relacionamento, criação do controller do usuário, criação do serviço campanha(ainda não terminado)

+++ src/main/java/br/ufpb/dcx/project/repository/RepositoryCampaign.java
@@ -3,5 +3,5 @@ package br.ufpb.dcx.project.repository;
 import br.ufpb.dcx.project.model.Campaign;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RepositoryCampaign extends JpaRepository<Campaign,Double> {
+public interface RepositoryCampaign extends JpaRepository<Campaign,Long> {
 }
